+++ docs/site/content/docs/remote-servers.mdx
@@ -129,9 +129,9 @@ Install Orca and its bundled CLI on the server, then run:
 <Callout title="On Linux, start it with orca-ide serve">
   The Linux CLI is named `orca-ide`, because GNOME Orca's screen reader already owns
   `/usr/bin/orca`. A packaged `orca serve` does write a bare `orca` into `~/.local/bin`, but only
-  while it is starting, so that shim can never be the command that starts the server. Read
-  `orca serve` as `orca-ide serve` throughout this page when the host is Linux. See
-  [Install → Linux](/docs/install#linux).
+  while it is starting, so that shim can never be the command that starts the server. Read `orca
+  serve` as `orca-ide serve` throughout this page when the host is Linux. See [Install →
+  Linux](/docs/install#linux).
 </Callout>
 
 ```bash
